Avoid cyclic import in binary format contract
Fixes #269
