+++ packages/runtime/src/binary-writer.ts
@@ -1,8 +1,21 @@
-import type {IBinaryWriter} from "./binary-format-contract";
-import type {WireType} from "./binary-format-contract";
+import type {BinaryWriteOptions, IBinaryWriter, WireType} from "./binary-format-contract";
 import {PbLong, PbULong} from "./pb-long";
 import {varint32write, varint64write} from "./goog-varint";
 import {assertFloat32, assertInt32, assertUInt32} from "./assert";
+
+
+const defaultsWrite: Readonly<BinaryWriteOptions> = {
+    writeUnknownFields: true,
+    writerFactory: () => new BinaryWriter(),
+};
+
+
+/**
+ * Make options for writing binary data form partial options.
+ */
+export function binaryWriteOptions(options?: Partial<BinaryWriteOptions>): Readonly<BinaryWriteOptions> {
+    return options ? {...defaultsWrite, ...options} : defaultsWrite;
+}
 
 
 /**
@@ -289,3 +302,4 @@ export class BinaryWriter implements IBinaryWriter {
 
 
 }
+
